Introduce driver model.
This commit refactors the code so that the actual networking can be based on other gems than Net::HTTP. The default driver is still  Net:HTTP based - this gem does not *need* any dependency outside of the ruby stdlib to work.

This also requires to introduce POD classes for responses (which was refactored to no longer need the Net::HTTP interface) and requests.

This work would allow to easily add alternative (e.g. faster) drivers, and would make sure that the existing test suite can be reused to validate these, alternative, implementations.