Refactor content type handling
After reviewing it turns out that the original_body attribute in the response doesn’t make sense. After all, the HTTP response has a body which *always* is to be interpreted with an encoding as per content headers. The old original_body was somehow sometimes enforced to be interpreted with some encoding, and later on - when getting a body - we reencoded that body again.

This doesn’t make sense. What we do instead now is to get rid of original_body, keep the body in the right encoding, and make sure that that body is properly encoded. No need for original_body.